fix(cv): scrub and float64 every block-mean CV read
One new static door beside `_pow2_clipped` -- `_finite_mean(cv,
default=0.0, axis=None)` -- and every LINEAR block-mean CV read in
numpy_backend.py now goes through it. It fixes two warts the ~30
hand-written copies each carried:

* float64. The buffers are float32 and `np.mean` of a float32 array
  accumulates in float32, so a CONSTANT CV of 0.3 read 0.29999998 over
  64 samples and 0.30000001 over 512: a held voltage on a jack was NOT
  block-size exact. Fifteen modules were affected and are now bit-exact
  at 64 vs 512 under a constant CV (filter mono + per-voice, crossover,
  sweep_eq, motion_eq, tilt_eq, loudness, compressor, reverb, flanger,
  phaser, lfo mono, slew mono + voice, wavetable_morph, bowed, wind).
* the min/max NaN wart. Python's min/max do not propagate NaN -- every
  comparison against NaN is False -- so `max(0.0, nan)` hands back 0.0
  and a "clamp" silently reads a NaN as a rail, while `min(max(nan, lo),
  hi)` propagates it instead: same idiom, two wrong answers, decided by
  argument order. np.clip passes NaN through and `int(round(nan))`
  raises ValueError out of the audio thread (bitcrusher.bits_cv). A
  non-finite mean is now read as "no modulation" BEFORE any clamp,
  int() or shape decision sees it, as _pow2_clipped already does for
  the octave exponents.

Call sites routed (39 edits):
  buffered/warping sink _governed_ratio.ratio_cv; supersaw.detune_cv
  (mono + per-voice); wavetable_morph.position_cv; filter.cutoff_cv x4
  and filter.resonance_cv x4 (mono, per-voice, voice-shared, fallback);
  function_generator _octaves + _per_slot (rate/rise/fall_cv);
  lfo.rate_cv (mono + voice); slew _oct (rise/fall_cv, both branches);
  crossover.freq_cv; motion_eq band{i}_freq_cv / _gain_cv / _q_cv;
  sweep_eq.freq_cv; loudness.level_cv; tilt_eq.tilt_cv;
  compressor.threshold_cv; reverb decay/damping/mix_cv; rotary.fast;
  flanger.rate_cv; phaser.rate_cv; pitch_shifter.pitch_cv;
  bitcrusher.bits_cv + rate_cv; sampler.pitch_cv; bowed.pitch_cv;
  wind.pitch_cv; drift.rate_cv; modal.pitch_cv.

What moved: 120 of the 156 shipped examples render BIT-EXACT (4 s,
block 512, seeded) -- and every example that moved patches a routed CV
port, none without. 34 of the 36 that moved are float32-ulp level (max
abs 3e-8 .. 6e-7, relative <= 8e-7). Two are chaotic models amplifying
an ulp: bowed_cello 5.2e-2 and wind_duet 1.0e-4, both starting at the
ulp floor in block 2 and growing over seconds. Swapping the helper's
reduction back to float32 restores ALL 36 bit-exactly, so nothing but
the mean's dtype moved. Both models are now block-size exact, which
they were not before (bowed 1.1e-5, wind 6.0e-6 between 64 and 512).

Left alone: the per-sample CV paths, the (V, F) -> (F,) voice collapses
in _input_buffer and friends (a different reduction), the meter/RMS and
window means, and the sibling-owned renderers (vowel, noise, clock,
clock_divider, freeze, cv_recorder, pluck, organ, tape, chorus).

tests/test_cv_scrub.py: 187 new tests -- the helper, the min/max wart
as an executable note, a constant-CV 64-vs-512 pin per module, the same
pin re-run against the old float32 reduction as a self-test (it must
fail there), NaN / +-1e6 renders finite, and a source tripwire counting
the door in each routed renderer. tests/test_filter.py: the per-voice
oracle now takes its own block mean in float64 -- it pins lfilter
against the per-sample loop, not the accumulation order, and a float32
reduction there put the two coefficient sets past the tolerance.

Suite: 4719 passed, 5 skipped.

+++ tests/test_filter.py
@@ -372,8 +372,16 @@ def _reference_filter_voice(backend, module, blocks, cv_blocks=None):
         )
         if per_voice:
             sr = backend.sample_rate
+            # float64 like the renderer (``_finite_mean``, 2026-09-22):
+            # this oracle exists to pin the lfilter path against the
+            # per-sample loop, not to restate how the block mean
+            # accumulates, and a float32 reduction here puts the two
+            # coefficient sets ~1e-7 apart -- which a resonant biquad
+            # amplifies past the tolerance.
             cpv = np.clip(
-                base * np.power(2.0, cv.mean(axis=1)), 20.0, sr * 0.45
+                base * np.power(2.0, cv.mean(axis=1, dtype=np.float64)),
+                20.0,
+                sr * 0.45,
             )
             qc = max(0.1, min(q, 20.0))
             w0 = 2.0 * np.pi * cpv / sr
